+++ packages/datasource-customizer/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @forestadmin/datasource-customizer [1.28.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.28.1...@forestadmin/datasource-customizer@1.28.2) (2023-09-13)
+
+
+### Bug Fixes
+
+* **datasource-customizer:** add hasFieldChanged function to handle change hook on this field ([#816](https://github.com/ForestAdmin/agent-nodejs/issues/816)) ([46bae6d](https://github.com/ForestAdmin/agent-nodejs/commit/46bae6d92d6fefe93b43a8a58ae6d7127fa3eacf))
+
+## @forestadmin/datasource-customizer [1.28.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.28.0...@forestadmin/datasource-customizer@1.28.1) (2023-09-13)
+
+
+### Bug Fixes
+
+* **datasource-customizer:** fix smart action default value behavior ([#817](https://github.com/ForestAdmin/agent-nodejs/issues/817)) ([2f9c4c1](https://github.com/ForestAdmin/agent-nodejs/commit/2f9c4c1a967bd4b2189ff10c24d536f4124f37a1))
+
 # @forestadmin/datasource-customizer [1.28.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.27.1...@forestadmin/datasource-customizer@1.28.0) (2023-09-12)