+++ packages/datasource-dummy/CHANGELOG.md
@@ -1,3 +1,23 @@
+## @forestadmin/datasource-dummy [1.0.61](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.60...@forestadmin/datasource-dummy@1.0.61) (2023-09-13)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.28.2
+
+## @forestadmin/datasource-dummy [1.0.60](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.59...@forestadmin/datasource-dummy@1.0.60) (2023-09-13)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.28.1
+
 ## @forestadmin/datasource-dummy [1.0.59](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.58...@forestadmin/datasource-dummy@1.0.59) (2023-09-12)